requirements.txt
# Alexandria Podcast Generation - Minimal Python Dependencies
# Install: pip install -r requirements.txt

# Core runtime
pydantic>=2.0.0
pyyaml>=6.0
numpy>=1.24.0

# API & CLI surfaces
fastapi>=0.104.0
httpx>=0.24.0
typer[all]>=0.9.0

# Testing stack
pytest>=7.4.0
pytest-asyncio>=0.21.0
hypothesis>=6.92.0
